extract tag filter stuff from desktop sidebar
