+++ src/app/domain/notes/components/tag-filter/tag-filter.component.html
@@ -1,0 +1,89 @@
+@if (hasTags()) {
+<div class="mb-6">
+  <button
+    (click)="onTagsToggle()"
+    class="flex items-center justify-between w-full px-4 py-3 text-gray-700 hover:text-[var(--tw-primary)] hover:bg-gray-50 rounded-xl transition-all duration-200"
+    [attr.aria-expanded]="isTagsExpanded()"
+    aria-controls="tags-list"
+  >
+    <div class="flex items-center gap-3">
+      <app-tag-icon [size]="20" [color]="'currentColor'" />
+      <span class="font-medium">Tags</span>
+      @if (selectedTags().length > 0) {
+      <span
+        class="text-xs bg-[var(--tw-primary-accent-bg)] text-[var(--tw-primary)] px-2 py-1 rounded-full"
+      >
+        {{ selectedTags().length }} selected
+      </span>
+      }
+    </div>
+    @if (isTagsExpanded()) {
+    <app-icon-chevron
+      [size]="16"
+      [color]="'currentColor'"
+      [orientation]="'down'"
+      class="transition-transform duration-300"
+    />
+    } @else {
+    <app-icon-chevron
+      [size]="16"
+      [color]="'currentColor'"
+      [orientation]="'right'"
+      class="transition-transform duration-300"
+    />
+    }
+  </button>
+
+  <div
+    id="tags-list"
+    class="overflow-hidden transition-all duration-300 ease-in-out"
+    [class]="
+      isTagsExpanded() ? 'max-h-80 opacity-100' : 'max-h-0 opacity-0'
+    "
+  >
+    <div class="px-4 py-2">
+      @if (availableTags().length === 0) {
+      <p class="text-sm text-gray-500 py-4 text-center">
+        No tags available
+      </p>
+      } @else {
+      <div class="max-h-60 overflow-y-auto space-y-2 mb-4">
+        @for (tag of availableTags(); track tag) {
+        <label
+          class="flex items-center gap-3 p-2 rounded-lg hover:bg-gray-50 cursor-pointer transition-colors duration-200"
+        >
+          <input
+            type="checkbox"
+            [checked]="isTagSelected(tag)"
+            (change)="onTagToggle(tag)"
+            class="w-4 h-4 text-[var(--tw-primary)] border-gray-300 rounded focus:ring-[var(--tw-primary)] focus:ring-2"
+            [attr.aria-label]="'Filter by ' + tag + ' tag'"
+          />
+          <span
+            class="flex-1 px-3 py-1 rounded-full text-sm transition-all duration-200"
+            [class]="
+              isTagSelected(tag)
+                ? 'bg-slate-200 text-slate-800 font-medium'
+                : 'bg-gray-100 text-gray-700 hover:bg-gray-200'
+            "
+          >
+            #{{ tag }}
+          </span>
+        </label>
+        }
+      </div>
+
+      @if (selectedTags().length > 0) {
+      <button
+        (click)="onClearFilters()"
+        class="w-full px-3 py-2 text-sm text-red-600 hover:text-red-700 hover:bg-red-50 rounded-lg transition-colors duration-200 flex items-center justify-center gap-2"
+        aria-label="Clear all tag filters"
+      >
+        <app-x-icon [size]="16" [color]="'currentColor'" />
+        Clear Filters
+      </button>
+      } }
+    </div>
+  </div>
+</div>
+}
